pcie_integ: fix secure access control for ecam and mmio regions
By default PCIe integration control register allows only secure access
to the ECAM and MMIO regions and non-secure access has to be explicitly
enabled. This patch adds new member (allow_non_secure_access) to the
configuration structure to allow the config data to specify that
ecam and mmio regions are meant for secure access only or not. This
member has to be set to true to allow non-secure access to the ECAM and
MMIO regions. Also update the platform's configuration data to allow
non-secure access.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Iff0afaac938843855bc0cc4f3afa63c5fac73c01