cmn700: multiple scg with hnf coordinates selection
Each SCG/HTG covers an address range and this address range can
be made to target a group of HN-Fs. These groups of HN-Fs are typically
bound by an arbitrary rectangle/square in the mesh. To aid automatic
programming of the HN-Fs in SCG/HTG along with the discovery process,
each SCG/HTG takes hnf_pos_start and hnf_pos_end in the config data.
HN-F nodes which are bound by this range will be assigned to the
respective SCG/HTG. This eliminates the process of manually looking at
the mesh and assigning the HN-F node ids to a SCG/HTG. To adapt to this
new change, update rdn2's cmn700 configuration to have one SCG region
which covers the entire mesh.

                                       hnf_pos_end
                                            xx
                                           xx
                                          xx
                   ┌─────────────────────xx
                   │                     │
                   │                     │
                   │                     │
                   │                     │
                   │    nth- SCG/HTG     │
                   │                     │
                   │                     │
                   │                     │
                   │                     │
                  xx─────────────────────┘
                 xx
                xx
               xx
        hnf_pos_start

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I08689dc44b98c0d648e31cb6db3b5c410eb2309d