fwk_module: Simplify expects & traps
Currently we perform a double check on some conditions like NULL
pointers etc. That is, we set an expectation and if that fails we trap.
However, given the way that fwk_expect() is implemented, it gives us
little advantage to call fwk_expect at all before a fwk_trap.

This patch replaces the expectation with a simple if statement,
improving readability and code size.

Change-Id: Iebf7760f3eb1ae128d4750266a3dfb0c2e8a552d
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>