sensor: add support for timestamp configuration
This patch adds timestamp api to enable or disable and configuring
it.
It is also added BS_FIRMWARE_HAS_SENSOR_TIMESTAMP flag in Makefile
or SCP_ENABLE_SENSOR_TIMESTAMP in cmake to enable or disable sensor
timestamp support. When SCMI Sensor V2 protocol is enabled this
feature is enabled automatically.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I63743dc2101ca9a1ef9ff9bb9e8a434198dbab43