sensors: Extended Attributes Support
This patch adds extended attributes support for the
sensor module.

The purpose of this patch is to add the fields associated
with the extended attributes to the senor HAL, which will
be populated by the sensor driver.

This patch will also include the changes for the build
system flags to enable the extended attributes.

The extended attributes flag is used to say if the sensor
will be populating extra fields related to the limits
of the sensor. These limits are included in those fields
added within this patch.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: I26d717288cc2a8c0725f81e6b4ba4ed6d2287ffd