module/smt: allow smt mailbox init without power domain module
This patch enables smt mailbox to be initialized without the power
domain module included in the firmware. This can be used by mcp ramfw
which does not include power domain module but uses smt module. Hence
BUILD_HAS_MOD_POWER_DOMAIN macro is used to conditionally compile the
code that subscribes to power domain state transition notifications
and initializes the mailbox based on the notification from power domain
module. Also add a condition such that the shared mailbox can be
initialized by the completer type smt channels only.

If the power domain module is not included in the firmware build then
the smt module initializes the smt mailbox during the start phase.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I56531f635c55f1414be548594778f0c9c6a2deb9