scmi: Shrink log strings
For some platforms the SCMI service name is longer than others and this
is reflected in trimmed logs, losing the tail of the error messages.
While each platform has the choice to set their own FMW_LOG_COLUMNS,
it is regardless efficient to keep logs as short as possible.

This patch is shrinking some SCMI error messages without affecting the
intended meaning of the logs.

Change-Id: I5926117efdafe9b21947ae9e70258c3e41939ad4
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>