scmi_perf: Enforce plugin handler requirements at build for cmake
The plugin handler required SCMI FastChannel to run. To make this
requirement explicit, this patch adds a check during build stage to
make sure that FastChannels are enabled if the plugin handler extension
is enabled. If the requirement is not met, the build system reports an
error.

Note that this is provided for cmake build system only at this stage.

At the same time, fix the header inclusions to correctly build the
plugin handler with cmake.

Change-Id: Ibc43882abd6142942b2eb7d11cc95df7465cdb9e
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>