products: Disable multi-threading for all products
This change is a precursor to the subsequent changes where
multi-threading support will be removed from the SCP firmware.

The change is inteded to allow platform owners to test respective
platforms without multi-threading.

The term multi-threaded mode is misleading since it implies multiple
threads of execution are concurrently supported
(like in an OS), but that is not true. Also multi-threading add call
overhead and could make things slower, and creates a larger memory
footprint. Simply all of above is not needed at present, hence we
expect multi-thread mode to be removed in near future from the firmware.

Change-Id: Ia5009ddf75a5f385ee8a8966e178900e462e6cf9
Signed-off-by: Girish Pathak <girish.pathak@arm.com>