module: publicise cmn_skeena as a generic scp module
This patch adds cmn_skeena module - the CMN interconnect in
Morello - to the core SCP modules group. CMN Skeena is based
on CMN600 but has additional changes to facilitate transportation
of the tag bits, 5x2 CMN mesh size, etc.

The Morello product will now be using this newly added core module
as opposed to having it as a private product-specific module.

Signed-off-by: Anurag Koul <anurag.koul@arm.com>
Co-authored-by: Manoj Kumar <manoj.kumar3@arm.com>
Co-authored-by: Chandni Cherukuri <chandni.cherukuri@arm.com>
Change-Id: Icb5e4cde65a74c407370da7f68dcecf3c1b76a83