cmake: set SCP_LOG_LEVEL only if it is not defined
SCP_LOG_LEVEL is used to define the minimum logging level for the
firmware. Currently, the framework's CMakeLists sets it to INFO for
Debug and WARN for release builds. It does not honor SCP_LOG_LEVEL
variable defined through compiler definition or through another
CMakeLists file, which locks SCP_LOG_LEVEL either to INFO or WARN. Fix
this by setting the SCP_LOG_LEVEL only if it is not defined already.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I767fd7e126359c38ef8fccd62c571d048c3b3f65