n1sdp_c2c: Request response for power domain set state
The secondary SCP must wait for power state transition to complete before
sending the response back to the primary over i2c. This patch changes
the set_state parameter to request response when the power state
transition is completed. Also, a new event handling is introduced to
check the response from power domain.

Signed-off-by: Tarek El-Sherbiny <tarek.el-sherbiny@arm.com>
Change-Id: Ib282e85ba14b5119edd84e00cf92f9a063962ad7