morello: Add SCP to PCC communication module
This is a simple protocol for sending and receiving messages between
SCP and PCC using shared memory. The buffer sizes must be defined to
be identical in size. Payload sizes must be identical as well.

The protocol works by having two FIFO-style buffers in shared memory,
each processors TX buffer is mapped as the others RX buffer and vice
versa. Since it's a simple FIFO implementation, no mutexes or other
access control methods are necessary as for a given buffer, there is
only one producer and one consumer. This guarantees that messages are
processed in the order that they are received.

Signed-off-by: Chandni Cherukuri <chandni.cherukuri@arm.com>
Signed-off-by: Anurag Koul <anurag.koul@arm.com>
Change-Id: I73bc46dd9db2eecd183b5c13e29d9e034fc76da8