ci/cmake: Enable Cppcheck
Cppcheck was disabled for CMake in ci scripts.
A couple of changes are done in order to enable it and make it
work properly. These changes are listed below.
Suppress list is updated by a wild card character for each
entry in the list to let Cppcheck parse the suppress list correctly
when used in CMake.
CMake build is forced to stop if Cppcheck reports any warnings or
errors.
Cppcheck is enabled in ci_cmake.py build script.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: I5d207a04698853e6991dc12df0997879c4d244d3