tc0: Add support for Build Variants
This patch is introducing a variant mechanism for TC0.
By design, TC0 has only a single configuration in terms of features
that it can support. At the same time, TC0 is a flexible platforms that
can be used to experiment features.

With the recent module additions like MPMM, Traffic-cop and Thermal
Management, it would be handy to have the option to add those features
into a platform, namely TC0, in one simple command, rather than having to
change the firmware builds manually.

With this in mind, we introduce two variants for TC0:
- TC0_VARIANT_STD
- TC0_VAR_EXPERIMENT_POWER

The former is the default variant and nothing will change.
With the latter, the build system adds all the dependencies required to
build a set of features that can be added later.

With this patch, Traffic-cop is built-in with the new variant rather
than being included if PERF_PLUGINS_HANDLER is chosen.

To use the new variant, simply append
EXTRA_CONFIG_ARGS+=-DSCP_PLATFORM_VARIANT=<VARIANT-IDX>
in the CMake build.

Please note that the variant mechanism is available only for CMake builds.

Change-Id: I824284828167a80f0054b80345c2774c8d85522d
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>