tc0: Add support for sensors for VARIANT 1
This patch adds support for sensors within the variant
TC0_VAR_EXPERIMENT_POWER.
This is required for the upcoming addition of the Thermal
Management which will require a temperature sensor.

Note that the sensor added with this patch is a fake one
and serves the purpose to have a complete s/w reference
support.

Change-Id: Ia42f68aedd82f901ecfb27c85a320ce6069f8c3d
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>