cmn: refactor "master/slave" to "manager"
Arm is committed to making the language we use inclusive, meaningful, and
respectful. This commit is part of the inclusive language refactoring series
where we replace offensive terminology with more inclusive ones.

As part of this change, the documentation alongside the
'struct cmn<version>_cfgm_reg' structs is fixed to state that they
belong to the master/manager, rather than the slave as erroneously
stated previously.

Signed-off-by: Brett Warren <brett.warren@arm.com>
Change-Id: I2700fba32b12f6a8e7495c55659a8cb82fa02ddc