product/tc0: Fix Voltage Domain configuration
The default voltage value was configured outside the configured
valid values range. This value is changed in this patch from 5000000
to 500000.

Additionally the element name in the configuration is changed from
"USB" to "DUMMY_VOLTD" and the corresponding identifiers are changed
from "*_USB" to "*_DUMMY". This is to avoid confusion and to emphasize
that this is an example configuration

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: I5269ec371d538984388e880341bf93b24e691315