product/rcar: Remove multi-threading use
This change is a precursor to the subsequent changes where
multi-threading support will be removed from the SCP firmware.

The term multi-threaded mode is misleading since it implies multiple
threads of execution are concurrently supported
(like in an OS), but that is not true. Also multi-threading add call
overhead and could make things slower, and creates a larger memory
footprint. Simply all of above is not needed at present, hence we
expect multi-thread mode to be removed in near future from the firmware

Change-Id: I00a0c882ae284de2350f4f2a12a6c98bbb34f827
Signed-off-by: Girish Pathak <girish.pathak@arm.com>