ci/cmake: rework `ci_cmake.py` script
This patch reworks `ci_cmake.py` in order to make it more flexible
for different configuration types and also scalable for
upcoming new ones. It moves the product configuration into a
new class that handles all possible definitions and also sets
default values.
It allows one to set special arguments for each configuration.

TC1 is included in the supported products for testing.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I5026e80a3bb23979c14b6e29621258519932a6eb