product/juno: Remove multi-threading use
This change is a precursor to the subsequent changes where
multi-threading support will be removed from the SCP firmware.

The term multi-threaded mode is misleading since it implies multiple
threads of execution are concurrently supported
(like in an OS), but that is not true. Also multi-threading add call
overhead and could make things slower, and creates a larger memory
footprint. Simply all of above is not needed at present, hence we
expect multi-thread mode to be removed in near future from the firmware

Change-Id: I5f971ee23d7f19965a1ba5c25a1c5d6508424853
Signed-off-by: Girish Pathak <girish.pathak@arm.com>