product/rdn2: publish isolated CPU list via SDS
Add support to publish a list of application CPUs that should not be
used on the platform. The MPID of such CPUs are put together into a
isolated CPU list and published via SDS. Typical reasons for isolating
certain CPUs include faulty CPUs being identified or restricting the
number of CPUs available for use on the platform.

The isolated CPUs list is supplied from the platform system module's
config data. This list is then published via the SDS to allow the next
stages of boot software to obtain the list of CPUs to isolate.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: I6f7d1045ea98ad4f33ae13bc6014dfd0768692c6