mod/scmi: Add handling for SCMI requester response
The current SCMI module only provides handling mechanism for
commands received from SCMI requesters. However there is no handling
mechanism for the responses from SCMI completers.

With this patch a check is done on the received SCMI message to
detect if it is a requester's command or a completer's response and
eventually call the corresponding protocol's handler.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: I2d39730a432bc38b14b8cc1732c049f6ae3367b0