scmi_pd:Add power domain set state sync request handling
This patch adds support for the sync mode set state request.
When the sync request is received by SCMI power domain it will
send an event to itself to switch the context. Then it will request
the new state from HAL. The SCMI PD will handle wait for the response
from HAL before terminating the command by responding to the SCMI agent.

Signed-off-by: Tarek El-Sherbiny <tarek.el-sherbiny@arm.com>
Change-Id: Ib2f8ab80b47654153f5eafe2818aad69147b574c