cppcheck: cppcheck output in verbose mode
CMake build executes cppcheck and compiler in parallel
In case of error, it's not very obvious from the log which
tool caused the corresponding error. This change introduces
a python script wrapper around cppcheck to display more
verbose messages when cppcheck is invoked and highlights
the cppcheck error.

This commit also updates cppcheck_suppress_list to suppress
false errors in fwk_log.c for lines 187 and 189.

Change-Id: Id83d5264410b3a2c931661b424281c435f091040
Signed-off-by: Girish Pathak <girish.pathak@arm.com>