system_power: respond with correct status for shutdown/reboot
The SCMI system power command for shutdown and reset is not responded
after the event has been successfully dispatched. But in the case of a
warm reset implementation in which the SCP core/subsystem does not reset
(only the application cores reset and reboot), SCP has to respond to the
last SCMI command, failing which SCP cannot handle subsequent SCMI
commands that the application cores send after the warm reboot sequence
is complete.

To fix this, ensure that the SCMI system power command for shutdown,
cold reset and warm reset is responded unconditionally.

Change-Id: I496daeb677166d657995127b7083741f6fc9af46
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>