power_domain: Add warm reset notifications
On a request for warm reset, send out a notification to any registered
modules that can handle the warm reset request. Typically, warm reset
would be platform specific implementation and so it is expected that
platform support module would register itself to be notified about the
warm reset request.

In addition to these changes, fix coding style issues that are not
directly related to the code being introduced in this patch

Change-Id: Ie7095b6accc4fb0e2014f4a812bcb79004f5cc4b
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>