sensor: Single value for trip points
Each trip point is expressed into two values, high and low,
and both combined provide the trip point (64bit).
However this is an extra effort as the comparisons in the module
are performed with the 64 bit value returned by the driver.

Simplify the trip point value by using a single 64bit instead of
splitting and re-combining the two halves.

Change-Id: I7c0e559682081d5ad669a58bedd34e875bfbcc02
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>