check_copyright: Check copyright on changed files only
The current copyright checker parses all the files in the repository
and mandates that all the files have to have up-to-date copyright.
This is not ideal since it imposes that every file must have the
current year's date even though no changes were applied in the same
year.

Change the assumption for the verification such that only modified
files need to have up-to-date copyright.

Change-Id: I897a42135f73f985461da34535cc3acd9944e9bc
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>