cmn700: fix ccla node type
CCLA node is internal to the mesh hierarchy. Fix the discovery and
configure routines to treat the CCLA node as an internal node instead of
returning device error if it is not an external node. In addition to
that, verify if the number of CCG_RA, CCG_HA and CCLA are equal.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Ic1cdd4d510e8c13231017e752cd2a928e4299f07