module: add scmi_sensor_req
This patch introduces a new module scmi_sensor_req.
The main functionality of this module is to send SCMI commands
to another entities that hold sensor values and to receive this data.

This module's interface is compatible with the sensor hal. Where it
can be used as a sensor driver. This way the module requesting sensor
data doesn't need to know if it gets it locally from a sensor driver
or externally from a different entity.

Only get_value function of the API is implemented.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: Ifa7a5a5b73346c8659c034e51121f9063da5b9ea