scmi_perf: Fix identifier type for domain_attributes_handler
When this handler executes and all the checks have completed, the
payload is formed to respond to the calling agent.
The payload also contains the domain's name.

The domain's name is taken by the configuration and
`fwk_module_get_element_name` is called. This function checks whether
the identifier's type passed as parameter is an element type.

When the Plugins Hanlder is built-in, the identifier passed to the
above function is a sub-element and not an element. The result is that
the name is not copied in the payload.

Fix the above by making sure the identifier's type provided to retrieve
the name is element and not sub_element.

Change-Id: I540b66caf2d3fc2a08bca690a4491822f25dbfd9
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>