cmn600: print cmn600 revision number
Add get_cmn600_revision_name() to convert the cmn600 revision number to
a matching string in TRM. During cmn600 discovery process, this revision
string is logged.

The definitions of the peripheral ID revision numbers have been updated.
The lower 4-bits of these definitions are the same for all revisions and
so these bits are discarded and converted it to an enum. This also
allows better indexing into the revision string array as well.

Signed-off-by: Tony K Nadackal <tony.nadackal@arm.com>
Change-Id: If59c40d14ff5c569c340cd4bbde1deda402856f7