module/smt: add new handler for requester type smt channels
Add new handler for requester type smt channels and refactor requester
and completer smt channel handlers.

The requester type smt channels send requests and receive the response.
Hence, requester type smt channels assume the role of caller and the
completer type smt channels are considered to be the callee.

Also, add new parameter in smt_transmit() function to determine if
acknowledgement interrupt is required for the message that we send.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: Ib34e35d8ed239f543573cd17c3b933a437bc5fde