debug: set BUILD_HAS_DEBUGGER compile definition
The affected platforms did not set BUILD_HAS_DEBUGGER in cmake,
despite having this in their respective make files. For sgi575 and
rdn1e1 this broke rd_alarm_idx.h. Other platforms don't have a
use in the source code currently for the flag, but it is added
regardless in case it is needed in the future.

Signed-off-by: Brett Warren <brett.warren@arm.com>
Change-Id: Ib1a192bd912b8ad315a37be06713c511b2dd60c2