module/smt: enable scmi module to release smt channel lock
Add a new function smt_release_channel_lock() that will set the smt
channel context lock to false. This function is intended to be used by
the scmi module when it has received and processed a response message
which does not require a further response and hence smt_respond() won't
be called. This will lead to channel context lock not being released
since smt_respond() doesn't get called. Hence add a function to release
the channel lock without having to respond back.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I0b6e3006c105398f40a918ae99c54bf16d9a78d4