module/scmi: add response message handler
Add a new function to handle the completion of response message
processing by protocol modules. This function will in-turn call the
release_channel_lock function from the mod_scmi_to_transport api.

This will be used in a scenario where the scmi module has received a
response message for a previously initiated SCMI message. It is assumed
that this response message does not require a further response. So once
the corresponding protocol handles the response message it can call the
response_message_handler() from scmi module to mark the completion.

Without this, we won't be able to process a subsequent scmi message on
the same transport channel since channel context lock would not have
been released for the message for which smt_respond() was not called.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: Ibcba5d114ac0f0ad809f12feb2173f6894fd1a54