sensor: typedef included to give signed or unsigned value options
The changes in this patch are to introduce a typedef, which will
allow either signed or unsigned sensor values. This new typedef is
used to be able to conditionally support int64, allowing the
introduction of negative value sensors.

It will be automatically enabled with SCMI-sensor-v2, however it
can also be enabled on its own using with the cmake tag
SCP_ENABLE_SENSOR_SIGNED_VALUE.

Existing platforms that cannot handle signed values will return
FWK_E_SUPPORT when the above build options are set.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: I0f24e31d4bc3861b1724d814e866ef97f09c4b88