morello: Enforce strict ordering for mailbox memory
The mailbox memory is shared between the SCP and the
OSPM agent, and is used for SCMI communication between
the two. The underlying memory region, which the
mailbox memory is mapped to, is currently configured as
Normal Memory (weak ordering) via the SCP MPU attributes.

This has been traced down to be the origin of the occasional
mailbox channel ownership errors as seen in the SCP FW logs -
the root cause being that on some occasions, the firmware
ends up reading the stale value of the channel ownership status
bit (as opposed to the reading value updated by the SCMI agent
during the SCMI comms) because of the reordering/buffering at
the interconnect.

Configuring the memory region to Strongly Ordered enforces a
strict ordering and gets rid of the mailbox channel ownership
errors.

Signed-off-by: Anurag Koul <anurag.koul@arm.com>
Change-Id: I69b2f74e111dde1b7d1e0a22ba85f101366eb86d