scmi: SCMI System Power requestor
The purpose of this patch is to introduce the System Power
Requestor module.

At this time, it is designed to work with a single element
and is being utilised by the SCMI System Power module.

The module takes an SCMI System Power command and forwards it
to its element, in such a way that it is received by them as if
it were purely a System Power command.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: I6968c46de430de8389c0ad773c4fe560446e2b6e