ci: Print build output/information
The CI is indicating whether the build failed or passed,
but is not printing the output of the build, which is
making it difficult to debug any problems that arise when
running the cmake-ci tests.

Generate files with the information of each build's output.
Store those files in the path given by the parameter -bod
(build output directory) or under the default folder
/tmp/scp/build-output.

Signed-off-by: Tomás Agustín González Orlando <tomasagustin.gonzalezorlando@arm.com>
Change-Id: I1b45cee047a23e54e4fe2f24414001b583c706bc