pl011: Simplify mod_pl011_init_ctx()
mod_pl011_init_ctx takes as only parameter a pointer to the
module context. That is owned by the module itself and locally
available. There is no need to provide the function a pointer
to it.
Also, the function now only returns successfully, so it can be
simplified even further.

Simplify the function by using direct access to the local
variable instead via its pointer and change the return value to
void.

Change-Id: Ib3b34658dc88cd672e24c61ea3fd1a5804db9198
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>