FWU: synquacer: Add FWU Multi Bank Update and BL2 boot support
Add FWU Multi Bank Update and boot BL2 from FIP image support for
SynQuacer platform.

When the DSW 3-4 is on, SynQuacer platform will decode FWU metadata
and platform metadata to choose the active bank of the FIP image
and boot BL2 from it.

This also enables the platform trial boot, which count up the boot
counter in the platform metadata, and when it reaches to the limit
(== 3) it switch back to the previous active bank. This counter will
stop when the BL33 (typically U-Boot) clears the counter.

Signed-off-by: Jassi Brar <jaswinder.singh@linaro.org>
Signed-off-by: Masami Hiramatsu <masami.hiramatsu@linaro.org>
Signed-off-by: Masahisa Kojima <masahisa.kojima@linaro.org>
Change-Id: I00b9c932fac5b4a9a231fcf2a050b766abefed61