ci: add build for each log level
The CI is only testing the build for the default log
level.
Add a parameter "ll" (log-level) to test the builds
with a specified log level.
This is done so that multiple runs of the ci_cmake
tests can be done in parallel by calling this script
multiple times with the different log levels.
These changes are done while improving the readability
of the constructor calls of Build() with the different optional
parameters.

Signed-off-by: Tomás Agustín González Orlando <tomasagustin.gonzalezorlando@arm.com>
Change-Id: I30004a65194618237643464b19337bafdde26610