module/mhu2: add support for transport module and in-band messages
Enable MHUv2 driver module to interface with transport HAL module and
add support for in-band messages. Also, remove the smt module
dependency. As part of the transition towards the transport module, the
smt module is supported by the mhu2, but it will be removed soon. In the
future SMT module will be removed from all the platforms and will be
replaced by the transport module

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I35c7fda56c436c420866e8fdb1c6367c54e3b711