module/transport: add support for firmware messages
Add interface to support firmware messages. This enables modules other
than scmi to use the transport module to send and receive firmware
related messages. The SCMI message support in the transport module has
been made optional, it will be included only if the scmi module is
included in the firmware. Also, add documentation for the transport
module.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: If7ea728d2c068717bbe5ce4658c24730862a54b8