morello: Replace SCP to PCC communication interface
The communication between SCP and the on-board PCC
(Platform Configuration Controller) on Morello, is
currently based on shared memory based transport -
which is prone to vulnerabilities and isn't secure.

This patch leverages the I2C interface between SCP and
the PCC to migrate the communication over to the I2C
channel.

Signed-off-by: Anurag Koul <anurag.koul@arm.com>
Change-Id: I33120c6e5699fa82349603ab8808059fe7b9dc6b