thermal-mgmt: add temperature protection functionality
This patch adds temperature protection functionality that
allows to configure two different alarms, warning and critical.
They can be configured independently but the critical level
threshold should be above the one for warning. Also when an alarm is
triggered it is reported to the driver through an API.

A common case of use is that the first threshold would be used
as first line of action against over temperature, while the
second could be used to initiate a sequence of hard power-off.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I76304abd2feafcfc15836870ea9f5b4cfd5cc550