juno/reset_domain: Fix conditional inclusion of notification_id
The inclusion of the notification_id member is now dependent on
whether or not BUILD_HAS_SCMI_NOTIFICATIONS build flag is set.
But in the reset_domain implementation, the functionality associated
with that identifier is conditional to the availability of framework
notifications BUILD_HAS_NOTIFICATION.

This creates a misalignment of build option for the correct use of
the reset domain features.

Replace BUILD_HAS_SCMI_NOTIFICATIONS with BUILD_HAS_NOTIFICATION
for the inclusion of the notification_id.

Change-Id: If596761faae827a9a7b3fe4101b36039b0fb2720
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>