morello: Enable voltage scaling in DVFS framework
So far, the Morello platform employed just the frequency
scaling as part of the DVFS framework. Now that the stable
voltage/frequency OPPs have been established for Morello,
add voltage scaling capability to the framework.

Morello has XR77128 as its PMIC in the product, so the
configuration changes required for the xr77128 and the
PSU module are accompanied here.

Co-authored-by: Himanshu Sharma <Himanshu.Sharma@arm.com>
Signed-off-by: Himanshu Sharma <Himanshu.Sharma@arm.com>
Signed-off-by: Anurag Koul <anurag.koul@arm.com>
Change-Id: I0d5e55a3402da68576b0b0bcacb2469325697014