product/rdn2: add support for rdn2 plattform variant 2
Add support for RD-N2 MultiChip patform variant. This variant is a quad
chip platform where the chips are connected via a high speed CCIX link.
Each chip has 4 clusters. The "PLATFORM_VARIANT" value is set to 2.

In addition to this, increase the chip address space to 64 TiB to
account for the new changes in the increase in the address bus width
in the rdn2 variants.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I01256b9b254fa656676f08715c53350fac8b1bcc