CMake/ArmClang: Add a check for CMAKE_SYSTEM_ARCH
CMake 3.20 and lower automatically maps the CMAKE_SYSTEM_PROCESSOR
variable and an undocumented CMAKE_SYSTEM_ARCH to compile and link
options for ARMClang. For example, the -mcpu=cortex-m55 flag is added
when CMAKE_SYSTEM_PROCESSOR is set to cortex-m55.

However specifiying extra CPU features such as setting
CMAKE_SYSTEM_PROCESSOR to "cortex-m55+nodsp" resulted in compilation
errors.

A workaround solution to this was introduced in CMake. The solution is
to use the specific ArmClang variable CMAKE_SYSTEM_ARCH instead of
CMAKE_SYSTEM_PROCESSOR which sets the -march flag instead and without
the restriction in CMAKE_SYSTEM_PROCESSOR. This way additional features
can be added.

more information can be found at:
https://gitlab.kitware.com/cmake/cmake/-/issues/21173
https://cmake.org/cmake/help/latest/policy/CMP0123.html

Differences between setting -march and -mcpu for cortex-m55 can be
found here in the Command-line options section:
https://community.arm.com/arm-community-blogs/b/architectures-and-
processors-blog/posts/armv8_2d00_m-based-processor-software-develo
pment-hints-and-tips

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: I30af983d5a12928acaf491e537f059e8f2ce1d61