module/optee_mhu: add notification support
Completer channel are used for notification. The client fills the channel with
a message that is then used to send back a notification.
Lock the thread with a mutex to wait until the notification is ready and
realease the mutex to unblock the thread that will then process the
notification.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>