unit_test: Add a new variable for the unit test file name
For modules that consist of more than one source code file,
there is a need to create multiple unit_test files for each c file.
This change introduces 2 new variables which need to be populated
by the local CMake file. By submitting this change it is possible
to have more than one unit_test file for the same module.

Existing modules have been modified to sync with this change.

Signed-off-by: Tarek El-Sherbiny <tarek.el-sherbiny@arm.com>
Change-Id: If588589a88b98f0d6b3d557bb6c4f3305ac63340