cmn600: add support for additional NON_HASH_TGT_NODEID registers
In latest revisions of cmn600, there are five NON_HASH_TGT_NODEID
registers. But last two NON_HASH_TGT_NODEID registers are implemented in
a different address offset.

So the cmn600 version is checked to calculate the maximum number of IO
memory regions supported and decide which NON_HASH_TGT_NODEID bank needs
to be programmed based on the group index.

Signed-off-by: Tony K Nadackal <tony.nadackal@arm.com>
Change-Id: Ib83c3fef5d5048eb3d24ec230da6c2f535b2eccc