scp: Fix fast channel compilation setup
Enable the BUILD_HAS_FAST_CHANNELS via the build system.

This constant may be used by any module that provides
fast channel functionality to other modules (hence the
generic constant name).
Currently, the only module that does so is the transport
module.

This commit also enables BUILD_HAS_FAST_CHANNELS in the case
where SCMI_PERF_FAST_CHANNELS is used.

In the future, if a new module that implements fast channel
functionality is created, then it should also make sure
that it enables the generic fast channels constant in the
same way as with SCMI_PERF_FAST_CHANNELS.

Change-Id: Ieac39cd97671456e3735b411af903d5a00b06e34
Signed-off-by: Tomás Agustín González Orlando <tomasagustin.gonzalezorlando@arm.com>