juno: Add support for platform variants
Add the option for platform variants when building for Juno. These
changes will allow the user to choose to build for `BOARD` or `FVP`.
BOARD is the default value. This will decrease the size of the built
binaries as FVP/BOARD specific components can be enabled/disabled
accordingly.

To specify the platform variant, append
PLATFORM_VARIANT=<`BOARD`/`FVP`>
to the build command.

Co-authored-by: mgtzim <mdgolamtanvir.zim@arm.com>
Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Signed-off-by: mgtzim <mdgolamtanvir.zim@arm.com>
Change-Id: I7e1741674dcb3ee69384bb86671ca402a593c7fc