cmn700: ccg: do not overwrite link control reg while enabling it
When SMP mode is enabled, CCRA and CCHA link control registers are
programmed with SMP mode enable bit. When enabling the CCRA and CCHA
links using the same link control registers, the existing values were
overwritten causing the SMP mode to be disabled. Fix this by doing
`bitwise OR` when setting the enable bit.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Ie582335bae055d913e095e5493193a9254efbffe