cmn700: ccg: add support for enabling CCLA to CCLA direct connect mode
CCG enables connecting the CXS interface from the CCLA on one CMN‑700 to
the CXS interface of the other. In this mode, the interfaces are
connected without going through the lower link layer and PHY controller
IP, essentially connecting the upper link layer (ULL) directly on both
sides CCLA. This mode is used in simulation environments for quick
system bringups. Add support to enable this mode. An option is also
added in `mod_cmn700_ccg_config` structure to enable this mode through
config data.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: If2164eb4740148a9c8b0b9e546b4d5fd70c9c25e