cmn700: refactor io and scg region programming to a separate function
Move the code to program IO region and System cache regions in RNSAM to
a separate function and move the RNSAM counters to the cmn700 context.

This is an incremental change, this new function will be used by the
functions to add and remove regions from RNSAM. This also reduces the
complexity of the cmn700_setup_sam function.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: Ic4d5b002bf730b8cbb80cbc0c65e928891b924af